requirements.txt
django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg[binary]==3.3.3
django-cors-headers==4.9.0
django-redis==6.0.0
drf-spectacular==0.29.0
python-dotenv
python-json-logger==3.3.0
sentry-sdk[django]==2.56.0
ruff==0.11.2
pytest==9.0.2
pytest-django==4.12.0
pytest-cov==6.1.0
